@@ -6,32 +6,31 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.19.6 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  get: ^4.6.1 # state managment getx
-  package_info_plus: ^1.4.0 # info package device
-  shared_preferences: ^2.0.13 # preferences
-  version: ^2.0.0 # compare versions
+  cupertino_icons: ^1.0.5
+  get: ^4.6.5 # state managment getx
+  package_info_plus: ^8.0.0 # info package device
+  shared_preferences: ^2.0.17 # preferences
+  version: ^3.0.2 # compare versions
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
-  mockito: ^5.1.0
-  build_runner: ^2.1.7
+  flutter_lints: ^3.0.2 
+  mockito: ^5.3.2
+  build_runner: ^2.3.3
 
 flutter:
 
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - assets/
 
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
