@@ -4,7 +4,7 @@ import 'package:template/src/routes.dart';
 import 'package:template/src/screens/auth/login_controller.dart';
 
 class LoginScreen extends GetView<LoginController> {
-  const LoginScreen({Key? key}) : super(key: key);
+  const LoginScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
